Crowbar updates and ERC fixes
This updates the Crowbar bus connectors and the layout of the
schematic to match the finalized pinouts. It also corrects an
error in the original schematic where the buses were not
properly routing due to lack of labels.

Finally, this change includes fixes for the ERC and adds the CPU
and RTC crystals to the system.